Revert "Merge pull request #35 in BIOS/agent-metric-store from ~E9928178/agent-metric-store-wip:master to master"
This reverts commit f61dc0047ddae9e54cf76cd0c90b2550ae65b25f, reversing
changes made to 9c6309ff91936d1b86a9d717fc61b4694f09743e.

+++ src/multi_row.cc
@@ -75,7 +75,7 @@ bool
     long now_ms = get_clock_ms();
     long elapsed_periodic_ms = (now_ms - _first_ms);
     //check if max time duration expired or max cache size limit reached
-    return (_row_cache.size()>=_max_row || elapsed_periodic_ms >= (long)_max_delay_s * 1000 );
+    return (_row_cache.size()>=_max_row || elapsed_periodic_ms >= _max_delay_s * 1000 );
     
 }
 /* return INSERT query or empty string if no value in cache available*/
